prune utils package
move typesContainSelectionSet to delegate package
remove typeContainsSelectionSet
